[Feature] Add multi-turn MDP and tool-use (#11)
* clean up dead code

* working on some idea

* working on refactor MDP env to support multi-step use case

* working on refactoring the mdp env class to support multi-step MDPs

* working on fix pg loss becomes zero after using new mdp env

* improve base mdp env convert to episode

* run more tests with custom chat template

* refactor DAPO code

* working on refactoring explore env

* working on explore env

* working on add basic tool-use to MDP Env

* refactoring logging and MDP env episode for tools

* update readme

* re-run tests on MATH

+++ src/rl4llm/core/training_mixin.py
@@ -602,7 +602,7 @@ class TrainingMixin:
 
         for i in range(batch_size):
             # Process single sample
-            sample_logits = logits[i, ...].float()
+            sample_logits = logits[i, ...]  # logits[i, ...].float()
             sample_logprobs_all = torch.log_softmax(sample_logits, dim=-1)
             sample_actions = actions[i, ...].unsqueeze(1)
             sample_logprob = torch.gather(
@@ -657,7 +657,7 @@ class TrainingMixin:
 
         for i in range(batch_size):
             # grab a single [seq_len, vocab_size] slice
-            sample_logits = logits[i].float()
+            sample_logits = logits[i]  # logits[i].float()
             # log‑softmax → [L, V]
             logp = torch.nn.functional.log_softmax(sample_logits, dim=-1)
             # p = exp(logp) → [L, V]
